[Deepanjan]Add . Feature to find Best Rated Hotel is added.

+++ src/Hotel.java
@@ -12,7 +12,7 @@ public class Hotel {
 	   public String hotelName;
 	   public int weekdayRate;
 	   public int weekendRate;
-	   public int totalCost;
+	   public long totalCost;
 	   public int hotelRating;
 	   
 
@@ -47,11 +47,11 @@ public class Hotel {
 	public void setWeekendRate(int weekendRate) {
 		this.weekendRate = weekendRate;
 	}
-	public int getTotalCost() {
+	public long getTotalCost() {
 		return totalCost;
 	}
 
-	public void setTotalCost(int totalCost) {
+	public void setTotalCost(long totalCost) {
 		this.totalCost = totalCost;
 	}
